feat(rdb-inventario): routed tabs en layout (Stock | Movimientos | Levantamientos)
Implementa el alcance v1 de la iniciativa `module-page-submodules`
(plan: docs/planning/module-page-submodules.md, ADR-005).

Patrón: módulos con sub-módulos exponen sus secciones como rutas
hermanas bajo un layout compartido que monta `<ModulePage>` +
`<ModuleHeader>` + `<RoutedModuleTabs>`. Tab activo derivado de
`usePathname()`. Sub-detalles profundos heredan el strip de tabs.

Cambios estructurales:

- `app/rdb/inventario/layout.tsx` (nuevo) — shell del módulo con
  3 tabs routed y `<RequireAccess>` levantado.
- `app/rdb/inventario/page.tsx` — reescrito como Stock view (default
  landing del módulo). Reduce de 1369 a ~450 líneas extrayendo helpers
  compartidos.
- `app/rdb/inventario/movimientos/page.tsx` (nuevo) — Movimientos view
  como ruta hermana de Stock.
- `app/rdb/inventario/levantamientos/page.tsx` — quita import + render
  de `<InventarioTabs>` (ahora los provee el layout) y wrapper
  `container max-w-6xl` para uniformidad con Stock/Movimientos.

Componentes nuevos:

- `components/module-page/routed-module-tabs.tsx` — `<RoutedModuleTabs>`
  Link-based con paridad visual a `<ModuleTabs>` (underline emerald-500),
  expone `aria-current="page"` + `role="tablist"`.
- `components/inventario/{types,utils,stock-detail-drawer,
   registrar-movimiento-dialog,print-stock-list}` — helpers extraídos
  del page.tsx grande para reuso entre Stock, Movimientos y futuros
  consumidores.

Cleanup:

- Borra `components/inventario/inventario-tabs.tsx` (deprecado por
  ADR-004 R1; ya sin consumidores tras este PR).
- Revierte la entrada `Levantamientos` del sidebar (deshace PR #203
  en `nav-config.ts`).

Validación local: `npm run typecheck` (0 errors), `npm run lint`
(0 errors, 57 warnings preexistentes), `npm run test:run` (161/161
passing), `prettier --check` clean.

Pendiente post-merge: smoke manual de Beto (refresh, browser
back/forward, responsive ≤375px, flujo Registrar Movimiento desde Stock).
